requirements_multimodal_clean.txt
absl-py==2.3.1
accelerate==1.12.0
annotated-types==0.7.0
astunparse==1.6.3
cachetools==6.2.4
cloudpickle==3.1.2
diffusers==0.36.0
filelock==3.20.2
flatbuffers==25.12.19
fsspec==2025.12.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
keras==2.15.0
libclang==18.1.1
lightning-utilities==0.15.2
llvmlite==0.46.0
lpips==0.1.4
Markdown==3.10
matplotlib==3.10.8
ml-dtypes==0.3.2
mpmath==1.3.0
munkres==1.1.4
networkx==3.6.1
numba==0.63.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
opencv-python==4.12.0
opencv-python-headless==4.12.0
opt_einsum==3.4.0
peft==0.18.1
protobuf==4.25.8
pyaml==25.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PySide6==6.10.1
regex==2025.11.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.7.0
scikit-optimize==0.10.2
sentry-sdk==2.49.0
shap==0.49.1
shellingham==1.5.4
shiboken6==6.10.1
slicer==0.0.8
smmap==5.0.2
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.9.1
torch-fidelity==0.3.0
torchmetrics==1.8.2
torchvision==0.24.1
transformers==4.57.6
triton==3.5.1
typer-slim==0.21.1
typing-inspection==0.4.2
wandb==0.24.0
Werkzeug==3.1.4
wrapt==1.14.2
xformers==0.0.33.post2
